feat: 투자예정지+체크리스트 dto생성

See merge request s11-bigdata-dist-sub1/S11P21B207!107

+++ back-end/tokime/src/main/java/com/ssafy/tokime/model/InvestmentPlannedLand.java
@@ -33,7 +33,7 @@ public class InvestmentPlannedLand {
     private String landGradient;
 
     @Column
-    private Long landPrice;
+    private Integer landPrice;
 
     @Column(length = 50)
     private String landRoad;
